% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alta_lake_210Pb}
\alias{alta_lake_210Pb}
\alias{alta_lake_drying}
\title{Alta Lake Gravity Core Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 32 columns.}
\usage{
alta_lake_210Pb

alta_lake_drying
}
\description{
Physical properties and lead-210 activity information from Alta Lake, Whistler,
British Columbia, Canada.
}
\references{
Dunnington DW, Spooner IS, White CE, et al (2016) A geochemical perspective on the impact of
development at Alta Lake, British Columbia, Canada. J Paleolimnol 56:315-330.
doi:10.1007/s10933-016-9919-x
}
\seealso{
\link[mudata2]{alta_lake}
}
\keyword{datasets}
