% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expect_ages_similar}
\alias{expect_ages_similar}
\title{Test age similarity}
\usage{
expect_ages_similar(
  calculated_ages,
  known_ages,
  max_delta = 1,
  age_range = 0:100,
  na.rm = FALSE
)
}
\arguments{
\item{calculated_ages}{The ages that we aren't sure about}

\item{known_ages}{The ages that we are sure about}

\item{max_delta}{The maximum error}

\item{age_range}{The age range we care about}

\item{na.rm}{Should missing ages be removed?}
}
\description{
The testing of this package requires a lot of testing whether or not
two sets of ages are similar. This function is designed for use in
this testing. The default is 1 year of error in the last 100 years.
}
\examples{
# similar ages
expect_ages_similar(1:10 + runif(10, -1, 1), 1:10, max_delta = 1)

# not similar ages
try(expect_ages_similar(1:10 + runif(10, -1, 1), 1:10, max_delta = 0.01))

}
