% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{finite_head_prop}
\alias{finite_head_prop}
\alias{finite_tail_prop}
\alias{finite_tail}
\alias{finite_head}
\title{Subset generators}
\usage{
finite_head_prop(x, y, prop_tail = 1)

finite_tail_prop(x, y, prop_head = 1)

finite_tail(x, y, n_tail = sum(is.finite(x) & is.finite(y)))

finite_head(x, y, n_head = sum(is.finite(x) & is.finite(y)))
}
\arguments{
\item{x}{An independent variable like depth or cumulative dry mass.}

\item{y}{A dependent variable that responds exponentially to \code{x}.}

\item{prop_tail, prop_head}{A minimum proportion of finite values
that should be included in the output}

\item{n_tail, n_head}{A minimum number of finite values
that should be included in the output}
}
\value{
A logical vector that can be used to subset \code{x} and \code{y}.
}
\description{
Subset generators
}
\examples{
x <- 1:5
y <- c(1, 2, NA, NA, 5)

finite_head(x, y, 2)
finite_tail(x, y, 2)
finite_head_prop(x, y, 0.5)
finite_tail_prop(x, y, 0.5)

}
