% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_finite}
\alias{max_finite}
\title{Max finite value}
\usage{
max_finite(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
The maximum finite value, disregarding  any
\code{\link[errors:errors]{errors::errors()}} that might be present.
}
\description{
Max finite value
}
