% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb210-package.R
\docType{package}
\name{pb210-package}
\alias{pb210}
\alias{pb210-package}
\title{pb210: Lead-210 dating utilities}
\description{
Utilities to create chronologies for sediment cores using
    analysis of lead-210 activity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/pb210}
  \item Report bugs at \url{https://github.com/paleolimbot/pb210/issues}
}

}
\author{
\strong{Maintainer}: 
    Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
