% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs-cic.R
\name{pb210_age_cic}
\alias{pb210_age_cic}
\alias{pb210_age_crs}
\title{Apply the Constant Rate of Supply or Constant Initial Concentration model}
\usage{
pb210_age_cic(depth, excess_pb210, excess_pb210_sd = NA_real_,
  calc_excess_pb210_surface = pb210_surface_min_depth,
  decay_constant = pb210_decay_constant())

pb210_age_crs(depth, excess_pb210, sample_mass,
  excess_pb210_sd = NA_real_,
  calc_excess_pb210_surface = pb210_surface_min_depth,
  calc_inventory_surface = pb210_surface_min_depth,
  calc_inventory_below = pb210_deep_inventory_zero,
  core_area = pb210_core_area(),
  decay_constant = pb210_decay_constant())
}
\arguments{
\item{depth}{The representative depth of each sample. Must be greater than 0.}

\item{excess_pb210}{Excess (non-erosional) lead-210 activity. Can be NA for samples on which
lead-210 was not measured.}

\item{excess_pb210_sd}{The standard deviation of the excess lead-210 value. Can be NA if error is not known.}

\item{calc_excess_pb210_surface, calc_inventory_surface, calc_inventory_below}{The surface
inventory and deep inventory (CRS) and surface lead-210 activity (CIC)
are important factors in the models that may need to be estimated. The default is to use the value
at the minimum depth value (closest to surface). purrr-style one-sided formulas are
supported.}

\item{decay_constant}{The decay contstant for lead-210, in 1/yr. This is an argument
because different spreadsheets use different decay constants (!!), and we would like
these functions to reproduce most existing age-depth models. Ideally this value should
be ln(2) / 22.26 (the default), given that the half-life of lead-210 is 22.26 years.}

\item{sample_mass}{The mass of each sample.}

\item{core_area}{The internal diameter of the corer, in units of length^2.}
}
\value{
A table with (at least) components \code{age} and \code{age_sd}. CRS model
output also contains \code{mar} and \code{mar_sd}.
}
\description{
Apply the Constant Rate of Supply or Constant Initial Concentration model
}
\references{
Appleby, P.G., and Oldfield, F. 1983. The assessment of 210Pb data from sites with
varying sediment accumulation rates.
Hydrobiologia, 103: 29–35. doi:10.1007/BF00028424.
}
