% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs-cic.R
\name{pb210_ages}
\alias{pb210_ages}
\alias{pb210_ages.pb210_fit_cic}
\alias{pb210_ages.pb210_fit_crs}
\title{Extract ages from fit objects}
\usage{
pb210_ages(object, cumulative_dry_mass = NULL)

\method{pb210_ages}{pb210_fit_cic}(object, cumulative_dry_mass = NULL)

\method{pb210_ages}{pb210_fit_crs}(object, cumulative_dry_mass = NULL)
}
\arguments{
\item{object}{A fit object generated by \code{\link[=pb210_crs]{pb210_crs()}} or \code{\link[=pb210_cic]{pb210_cic()}}}

\item{cumulative_dry_mass}{The cumulative dry mass of the core (in kg), starting at the
surface sample and including all samples in the core.
These must be greater than 0 and in increasing order.}
}
\value{
A table with (at least) components \code{age} and \code{age_sd} (both in years).
CRS model output also contains \code{mar} and \code{mar_sd} (in kg / m^2^ / year).
}
\description{
Extract ages from fit objects
}
