% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pb210_as_fit}
\alias{pb210_as_fit}
\alias{pb210_as_fit.default}
\alias{pb210_as_fit.numeric}
\title{Coerce objects to exponential fits}
\usage{
pb210_as_fit(x)

\method{pb210_as_fit}{default}(x)

\method{pb210_as_fit}{numeric}(x)
}
\arguments{
\item{x}{An object}
}
\value{
An object with a \code{\link[stats:predict]{stats::predict()}} method.
}
\description{
This is used to sanitize input for \code{\link[=pb210_age_cic]{pb210_age_cic()}} and \code{\link[=pb210_age_crs]{pb210_age_crs()}},
where fit objects are required but where it is anticipated that
numeric constants will be common as input.
}
\examples{
fake_depth <- 0:10
fake_pb210 <- exp(5 - fake_depth) + rnorm(11, sd = 0.005)

pb210_as_fit(pb210_fit_exponential(fake_depth, fake_pb210))
pb210_as_fit(1)
pb210_as_fit(0)

}
