% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs-cic.R
\name{pb210_calculate_inventory}
\alias{pb210_calculate_inventory}
\title{Calculate cumulative lead-210 activity}
\usage{
pb210_calculate_inventory(cumulative_dry_mass,
  excess_pb210_specific_activity,
  model_bottom = pb210_fit_exponential(cumulative_dry_mass,
  excess_pb210_specific_activity),
  model_middle = pb210_fit_interpolator_linear(cumulative_dry_mass,
  excess_pb210_specific_activity), n_segments = 200L)
}
\arguments{
\item{cumulative_dry_mass}{The cumulative dry mass of the core in kilograms, starting at depth
0. These must be positive and in increasing order.}

\item{excess_pb210_specific_activity}{An excess lead-210 activity for samples where this was
measured, and NA where lead-210 was not measured (NA values will be interpolated using
\code{model_middle}).}

\item{model_bottom}{A fit object that will be used to model activities below
the last positive finite lead-210 activity. This must be created using
\code{\link[=pb210_fit_exponential]{pb210_fit_exponential()}} in that its \code{m} and \code{b} coefficients are used to calculate
the integrated depth below the last positive finite \code{excess_pb210_specific_activity}.}

\item{model_middle}{A fit object that will be used to model activities between
measured finite lead-210 activities.}

\item{n_segments}{The number of tiny rectangles used to approximate the cumulative
activity between the first and last positive finite lead-210 measurement.}
}
\value{
A vector of cumulative lead-210 activities for each sample.
}
\description{
The cumulative content of lead-210 from the bottom of the core is the basis for the
constant rate of supply model, and is a required input to \code{\link[=pb210_age_crs]{pb210_age_crs()}}.
}
\examples{
fake_mass <- 1:10
fake_pb210 <- exp(5 - fake_mass) + rnorm(10, sd = 0.005)
pb210_calculate_inventory(fake_mass, fake_pb210)

}
