% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs-cic.R
\name{pb210_cic}
\alias{pb210_cic}
\alias{pb210_crs}
\alias{predict.pb210_fit_cic}
\alias{predict.pb210_fit_crs}
\title{Apply the Constant Rate of Supply or Constant Initial Concentration model}
\usage{
pb210_cic(cumulative_dry_mass, excess,
  model_top = ~pb210_fit_exponential(..1, ..2),
  decay_constant = pb210_decay_constant())

pb210_crs(cumulative_dry_mass, excess,
  inventory = pb210_inventory_calculator(),
  model_top = ~pb210_fit_exponential(..1, ..2),
  core_area = pb210_core_area(),
  decay_constant = pb210_decay_constant())

\method{predict}{pb210_fit_cic}(object, cumulative_dry_mass = NULL, ...)

\method{predict}{pb210_fit_crs}(object, cumulative_dry_mass = NULL, ...)
}
\arguments{
\item{cumulative_dry_mass}{The cumulative dry mass of the core (in kg), starting at the
surface sample and including all samples in the core.
These must be greater than 0 and in increasing order.}

\item{excess}{An excess (non-erosional) lead-210 specific activity (in Bq/kg)
for samples where this was measured, and NA where lead-210 was not measured. Use
\code{\link[errors:set_errors]{errors::set_errors()}} to use automatic error propogation.}

\item{model_top}{A fit object, such as one generated by \code{\link[=pb210_fit_exponential]{pb210_fit_exponential()}}, or a
constant specifying the surface concentration (CIC model) or inventory (CRS model).
The choice of this value has considerable impact on young dates.}

\item{decay_constant}{The decay contstant for lead-210 (in 1/years). This is an argument
rather than a constant because we have found that different spreadsheets in the wild
use different decay constants. See \code{\link[=pb210_decay_constant]{pb210_decay_constant()}}.}

\item{inventory}{The cumulative excess lead-210 activity (in Bq), starting at the bottom
of the core. By default, this is estimated by the default \code{\link[=pb210_inventory_calculator]{pb210_inventory_calculator()}}.}

\item{core_area}{The internal area of the corer (in m^2^). This can be calculated
from an internal diameter using \code{\link[=pb210_core_area]{pb210_core_area()}}.}

\item{object}{A fit object generated by \code{\link[=pb210_crs]{pb210_crs()}} or \code{\link[=pb210_cic]{pb210_cic()}}.}

\item{...}{Unused.}
}
\value{
\code{predict()} methods return a tibble with (at least)
components \code{age} and \code{age_sd} (both in years).
CRS model \code{predict()} function output also contains \code{inventory}, \code{inventory_sd},
\code{mar} and \code{mar_sd} (in kg / m^2^ / year).
}
\description{
Apply the Constant Rate of Supply or Constant Initial Concentration model
}
\examples{
# simulate a core
core <- pb210_simulate_accumulation(
  mass_accumulation = pb210_mass_accumulation_constant()
) \%>\%
  pb210_simulate_core() \%>\%
  pb210_simulate_counting()

# calculate ages using the CIC model
cic <- pb210_cic(
  pb210_cumulative_mass(core$slice_mass, position = 0.5),
  set_errors(
    core$pb210_specific_activity_estimate,
    core$pb210_specific_activity_se
  )
)

predict(cic)

# calculate ages using the CRS model
crs <- pb210_crs(
  pb210_cumulative_mass(core$slice_mass),
  set_errors(
    core$pb210_specific_activity_estimate,
    core$pb210_specific_activity_se
  )
)

predict(crs)

}
\references{
Appleby, P.G., and Oldfield, F. 1983. The assessment of ^210^Pb data from sites with
varying sediment accumulation rates.
Hydrobiologia, 103: 29–35. doi:10.1007/BF00028424.

Appleby, P.G., and Oldfield, F. 1978. The calculation of lead-210 dates assuming a
constant rate of supply of unsupported ^210^Pb to the sediment.
CATENA, 5: 1–8. doi:10.1016/S0341-8162(78)80002-2.
}
