% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{pb210_cic_monte_carlo}
\alias{pb210_cic_monte_carlo}
\alias{pb210_crs_monte_arlo}
\alias{predict.pb210_fit_cic_monte_carlo}
\title{Run a Monte-Carlo simulation on a CRS or CIC fit}
\usage{
pb210_cic_monte_carlo(
  cumulative_dry_mass,
  activity,
  background = 0,
  model_top = ~pb210_fit_exponential(..1, ..2),
  decay_constant = pb210_decay_constant(),
  n = 100,
  sample_activity = pb210_sample_norm,
  sample_background = pb210_sample_norm,
  sample_decay_constant = pb210_sample_norm
)

pb210_crs_monte_arlo(
  cumulative_dry_mass,
  excess,
  inventory = pb210_inventory_calculator(),
  core_area = pb210_core_area(),
  decay_constant = pb210_decay_constant()
)

\method{predict}{pb210_fit_cic_monte_carlo}(
  object,
  cumulative_dry_mass = NULL,
  ...,
  summarise_value = stats::median,
  summarise_sd = ~sd(..1)/sqrt(length(..1))
)
}
\arguments{
\item{cumulative_dry_mass}{The cumulative dry mass of the core (in kg), starting at the
surface sample and including all samples in the core.
These must be greater than 0 and in increasing order.}

\item{activity}{A vector of measured lead-210 specific activities (in Bq/kg) and
associated error. These can have \code{\link[errors:errors]{errors::errors()}}.}

\item{background}{A vector of estimated background
lead-210 specific activity (in Bq/kg) and associated error.
These can have \code{\link[errors:errors]{errors::errors()}}.}

\item{model_top}{A fit object, such as one generated by \code{\link[=pb210_fit_exponential]{pb210_fit_exponential()}} or a
constant specifying the surface \code{excess}. The choice of this value has
considerable impact on young dates.}

\item{decay_constant}{The decay contstant for lead-210 (in 1/years). This is an argument
rather than a constant because we have found that different spreadsheets in the wild
use different decay constants. See \code{\link[=pb210_decay_constant]{pb210_decay_constant()}}.}

\item{n}{The number of permutations. The default is 1,000, as
Sanchez-Cabeza et al. (2014) found that this was the minimum
number of iterations needed for Monte-Carlo uncertainty to
converge on the quadrature-propegated uncertainty. In general,
Sanchez-Cabeza et al. (2014) used n values from 1,000 to 4,000.}

\item{sample_activity, sample_background, sample_decay_constant}{Random
sampler functions such as \code{\link[=pb210_sample_norm]{pb210_sample_norm()}} that are called
\code{n} times for the appropriate argument.}

\item{excess}{An excess (non-erosional) lead-210 specific activity (in Bq/kg)
for samples where this was measured, and NA where lead-210 was not measured. Use
\code{\link[errors:set_errors]{errors::set_errors()}} to use automatic error propogation.}

\item{inventory}{The cumulative excess lead-210 activity (in Bq), starting at the bottom
of the core. By default, this is estimated by the default \code{\link[=pb210_inventory_calculator]{pb210_inventory_calculator()}}.
If specifying a vector of values, ensure  that the surface (0 cumulative mass) value is
specified.}

\item{core_area}{The internal area of the corer (in m^2^). This can be calculated
from an internal diameter using \code{\link[=pb210_core_area]{pb210_core_area()}}.}

\item{object}{A fit object generated by \code{\link[=pb210_crs]{pb210_crs()}} or \code{\link[=pb210_cic]{pb210_cic()}}.}

\item{...}{Unused.}

\item{summarise_value, summarise_sd}{Functions that accept a vector
of values and calculate a summary value or error, respectively.
The default calculates a median and standard error.}
}
\description{
Run a Monte-Carlo simulation on a CRS or CIC fit
}
\references{
Binford, M.W. 1990. Calculation and uncertainty analysis of ^210^Pb dates for
PIRLA project lake sediment cores. Journal of Paleolimnology, 3: 253–267.
doi:10.1007/BF00219461.

Sanchez-Cabeza, J.-A., Ruiz-Fernández, A.C., Ontiveros-Cuadras, J.F.,
Pérez Bernal, L.H., and Olid, C. 2014. Monte Carlo uncertainty calculation of ^210^Pb
chronologies and accumulation rates of sediments and peat bogs.
Quaternary Geochronology, 23: 80–93. doi:10.1016/j.quageo.2014.06.002.
}
