% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pb210_compressibility_constant}
\alias{pb210_compressibility_constant}
\alias{pb210_compressibility_none}
\title{Parameter generators for compressibility}
\usage{
pb210_compressibility_constant(value = 0.001)

pb210_compressibility_none()
}
\arguments{
\item{value}{A compressibility at the time of deposition, in 1/Pa.
Higher values (closer to 0) indicate a higher propensity for the material
to be compressed.}
}
\value{
A function of a single parameter, \code{age}, which is a (decreasing) vector of ages.
}
\description{
Parameter generators for compressibility
}
\examples{
ages <- 150:0
pb210_compressibility_constant()(ages)

}
