% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{pb210_crs}
\alias{pb210_crs}
\title{Apply the Constant Rate of Supply model}
\usage{
pb210_crs(cumulative_dry_mass, excess_pb210, pb210_sd = NA,
  internal_diameter = 6.3)
}
\arguments{
\item{cumulative_dry_mass}{The cumulative dry mass of the section top}

\item{excess_pb210}{Excess (non-erosional) lead-210 activity}

\item{pb210_sd}{The standard deviation of the excess lead-210 value}

\item{internal_diameter}{The internal diameter of the corer, in cm}
}
\value{
A model fit object
}
\description{
Apply the Constant Rate of Supply model
}
\references{
Appleby, P.G., and Oldfield, F. 1983. The assessment of 210Pb data from sites with
varying sediment accumulation rates.
Hydrobiologia, 103: 29–35. doi:10.1007/BF00028424.
}
