% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_parameters.R
\name{pb210_cumulative_mass}
\alias{pb210_cumulative_mass}
\title{Calculate cumulative mass}
\usage{
pb210_cumulative_mass(masses, position = 0)
}
\arguments{
\item{masses}{A vector of sequential masses}

\item{position}{0 for the cumulative mass at the bottom of the section
and 1 for the cumulative mass}
}
\value{
A vector of cumulative masses
}
\description{
Calculate cumulative mass
}
\examples{
# top of slice
pb210_cumulative_mass(c(0.1, 0.1, 0.1, 0.1), position = 0)

# middle of slice
pb210_cumulative_mass(c(0.1, 0.1, 0.1, 0.1), position = 0.5)

# bottom of slice
pb210_cumulative_mass(c(0.1, 0.1, 0.1, 0.1), position = 1)

}
