% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pb210_density_constant}
\alias{pb210_density_constant}
\title{Parameter generators for density values}
\usage{
pb210_density_constant(value = 1500)
}
\arguments{
\item{value}{A density at the time of deposition, in kg / m3}
}
\value{
A function of a single parameter, `age`, which is a (decreasing) vector of ages.
}
\description{
Note that this refers to dry density at the time of deposition.
}
\examples{
age_compare <- tibble::tibble(
  ages = 150:0,
  pb210_density_constant()(ages)
)

}
