% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pb210_fit_interpolator_linear}
\alias{pb210_fit_interpolator_linear}
\title{Fit an interpolator}
\usage{
pb210_fit_interpolator_linear(x, y)
}
\arguments{
\item{x, y}{Numeric vectors of known values, where \code{x} is the
"known" variable and \code{y} is the variable that must be interpolated
(between unknown \code{x} values).}
}
\value{
An object of class "interpolator".
}
\description{
Fit an interpolator
}
\examples{
fake_depth <- 0:10
fake_pb210 <- exp(5 - fake_depth) + rnorm(11, sd = 0.005)
fit_interp <- pb210_fit_interpolator_linear(fake_depth, fake_pb210)
predict(fit_interp, newdata = tibble::tibble(x = seq(-1, 11, by = 0.5)))

}
