% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{pb210_sample_norm}
\alias{pb210_sample_norm}
\title{Random samplers}
\usage{
pb210_sample_norm(x)
}
\arguments{
\item{x}{A vector of values, possibly with \code{\link[errors:errors]{errors::errors()}}.}
}
\value{
A (errorless) vector of randomly sampled values if \code{x} has
\code{\link[errors:errors]{errors::errors()}}, otherwise \code{x}, unchanged.
}
\description{
For use in \code{\link[=pb210_crs_monte_carlo]{pb210_crs_monte_carlo()}}. In the future, other distributions
may be considered; we approximate mean + standard error as a normal distribution
because this is done in
}
\examples{
pb210_sample_norm(set_errors(1:10, 1))
pb210_sample_norm(set_errors(1:10, 1))
pb210_sample_norm(1:10)

}
\references{
Aquino-López, M.A., Blaauw, M., Christen, J.A.,
and Sanderson, N.K. 2018. Bayesian Analysis of ^210^Pb Dating.
Journal of Agricultural, Biological and Environmental Statistics, 23: 317–333.
\url{https://doi.org/10.1007/s13253-018-0328-7}.
}
