% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pb210_supply_constant}
\alias{pb210_supply_constant}
\title{Parameter generators for lead-210 supply}
\usage{
pb210_supply_constant(value = 100)
}
\arguments{
\item{value}{An flux of lead 210, in Bq / m^2^ / year.}
}
\value{
A function of a single parameter, \code{age}, which is a (decreasing) vector of ages.
}
\description{
Parameter generators for lead-210 supply
}
\examples{
ages <- 150:0

}
\references{
Appleby, P.G. 2008. Three decades of dating recent sediments by fallout
radionuclides: a review. The Holocene, 18: 83–93. doi:10.1177/0959683607085598.
}
