% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs-cic.R
\name{pb210_surface_min_depth}
\alias{pb210_surface_min_depth}
\alias{pb210_surface_estimate}
\title{Estimate surface Lead-210 activity}
\usage{
pb210_surface_min_depth(depth, pb210)

pb210_surface_estimate(depth, pb210)
}
\arguments{
\item{depth}{Depth of sample}

\item{pb210}{Activity of lead-210 or inventory}
}
\value{
A plausible surface value given the inputs
}
\description{
Estimates surface lead-210 activity using an exponential fit of `pb210`
vs. `-depth`, or using the lead-210 activity at the  minimum `depth` value.
}
