% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pb210_time_seconds}
\alias{pb210_time_seconds}
\title{Generate a time duration in seconds}
\usage{
pb210_time_seconds(days = 0, hours = 0, minutes = 0, seconds = 0)
}
\arguments{
\item{days, hours, minutes, seconds}{Quantities to convert to seconds. Multiple inputs will be
added.}
}
\value{
A time in seconds
}
\description{
Generate a time duration in seconds
}
\examples{
pb210_time_seconds(days = 1)
pb210_time_seconds(hours = 24)

}
