% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pb210_water_content_constant}
\alias{pb210_water_content_constant}
\title{Parameter generators for water content}
\usage{
pb210_water_content_constant(value = 0.95)
}
\arguments{
\item{value}{A water content at the time of deposition, unitless (mass water / mass sample)}
}
\value{
A function of a single parameter, \code{age}, which is a (decreasing) vector of ages.
}
\description{
Note that this refers to water content at the time of deposition.
}
\examples{
ages <- 150:0
pb210_water_content_constant()(ages)

}
