% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb210-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{drop_errors}
\alias{set_errors}
\alias{errors}
\alias{predict}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{errors}{\code{\link[errors]{drop_errors}}, \code{\link[errors]{errors}}, \code{\link[errors]{set_errors}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{stats}{\code{\link[stats]{predict}}}
}}

