% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pb210_as_fit}
\alias{pb210_as_fit}
\alias{pb210_as_fit.default}
\alias{pb210_as_fit.numeric}
\alias{pb210_fit_lazy}
\alias{pb210_as_fit.function}
\alias{pb210_as_fit.formula}
\alias{pb210_as_fit.pb210_fit_lazy}
\title{Coerce objects to exponential fits}
\usage{
pb210_as_fit(x, ...)

\method{pb210_as_fit}{default}(x, ...)

\method{pb210_as_fit}{numeric}(x, ...)

pb210_fit_lazy(x)

\method{pb210_as_fit}{function}(x, ...)

\method{pb210_as_fit}{formula}(x, ...)

\method{pb210_as_fit}{pb210_fit_lazy}(x, data = NULL, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Unused.}

\item{data}{For lazy fits, this object is a tibble with
\code{cumulative_dry_mass} and \code{excess}.}
}
\value{
An object with a \code{\link[stats:predict]{stats::predict()}} method.
}
\description{
This is used to sanitize input for \code{\link[=pb210_cic]{pb210_cic()}} and \code{\link[=pb210_crs]{pb210_crs()}},
where fit objects are required but where it is anticipated that
numeric constants will be common as input.
}
\examples{
fake_depth <- 0:10
fake_pb210 <- exp(5 - fake_depth) + rnorm(11, sd = 0.005)

pb210_as_fit(pb210_fit_exponential(fake_depth, fake_pb210))
pb210_as_fit(1)
pb210_as_fit(0)

}
