% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs-cic.R
\name{pb210_cic}
\alias{pb210_cic}
\alias{pb210_crs}
\alias{pb210_age_cic}
\alias{pb210_age_crs}
\title{Apply the Constant Rate of Supply or Constant Initial Concentration model}
\usage{
pb210_cic(cumulative_dry_mass, excess_pb210, excess_pb210_sd = NA_real_,
  model_top = pb210_fit_exponential(cumulative_dry_mass, excess_pb210),
  decay_constant = pb210_decay_constant())

pb210_crs(cumulative_dry_mass, excess_pb210, excess_pb210_sd = NA_real_,
  inventory = pb210_inventory(cumulative_dry_mass, excess_pb210,
  excess_pb210_sd), inventory_sd = NA_real_,
  model_top = pb210_fit_exponential(cumulative_dry_mass, inventory),
  core_area = pb210_core_area(),
  decay_constant = pb210_decay_constant())

pb210_age_cic(cumulative_dry_mass, excess_pb210,
  excess_pb210_sd = NA_real_,
  model_top = pb210_fit_exponential(cumulative_dry_mass, excess_pb210),
  decay_constant = pb210_decay_constant())

pb210_age_crs(cumulative_dry_mass, excess_pb210,
  excess_pb210_sd = NA_real_,
  inventory = pb210_inventory(cumulative_dry_mass, excess_pb210,
  excess_pb210_sd),
  model_top = pb210_fit_exponential(cumulative_dry_mass, inventory),
  core_area = pb210_core_area(),
  decay_constant = pb210_decay_constant())
}
\arguments{
\item{cumulative_dry_mass}{The cumulative dry mass of the core (in kg), starting at the
surface sample and including all samples in the core.
These must be greater than 0 and in increasing order.}

\item{excess_pb210}{An excess (non-erosional) lead-210 specific activity (in Bq/kg)
for samples where this was measured, and NA where lead-210 was not measured.}

\item{excess_pb210_sd}{The standard error of the excess lead-210 specific activity (in Bq/kg).
Can be NA if error is not known.}

\item{model_top}{A fit object, such as one generated by \code{\link[=pb210_fit_exponential]{pb210_fit_exponential()}}, or a
constant specifying the surface concentration (CIC model) or inventory (CRS model).
The choice of this value has considerable impact on young dates.}

\item{decay_constant}{The decay contstant for lead-210 (in 1/years). This is an argument
rather than a constant because we have found that different spreadsheets in the wild
use different decay constants. See \code{\link[=pb210_decay_constant]{pb210_decay_constant()}}.}

\item{inventory}{The cumulative excess lead-210 activity (in Bq), starting at the bottom
of the core. By default, this is estimated by \code{\link[=pb210_inventory]{pb210_inventory()}}.}

\item{core_area}{The internal area of the corer (in m^2^). This can be calculated
from an internal diameter using \code{\link[=pb210_core_area]{pb210_core_area()}}.}
}
\value{
A table with (at least) components \code{age} and \code{age_sd} (both in years).
CRS model output also contains \code{mar} and \code{mar_sd} (in kg / m^2^ / year).
}
\description{
Apply the Constant Rate of Supply or Constant Initial Concentration model
}
\references{
Appleby, P.G., and Oldfield, F. 1983. The assessment of ^210^Pb data from sites with
varying sediment accumulation rates.
Hydrobiologia, 103: 29–35. doi:10.1007/BF00028424.

Appleby, P.G., and Oldfield, F. 1978. The calculation of lead-210 dates assuming a
constant rate of supply of unsupported ^210^Pb to the sediment.
CATENA, 5: 1–8. doi:10.1016/S0341-8162(78)80002-2.
}
