% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pb210_compressibility_constant}
\alias{pb210_compressibility_constant}
\title{Parameter generators for compressibility}
\usage{
pb210_compressibility_constant(value = 2e-06)
}
\arguments{
\item{value}{A compressibility at the time of deposition, in 1/Pa.
Lower values indicate a higher propensity for the material to be compressed.}
}
\value{
A function of a single parameter, \code{age}, which is a (decreasing) vector of ages.
}
\description{
Note that this refers to compressibility at the time of deposition.
}
\examples{
ages <- 150:0
pb210_compressibility_constant()(ages)

}
