% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pb210_core_area}
\alias{pb210_core_area}
\title{Calculate core area from an internal diameter}
\usage{
pb210_core_area(diameter = 0.063)
}
\arguments{
\item{diameter}{A diameter in meters. The default is 0.063 m (6.3 cm), which
is the internal diameter of the core tubes for our trusty 2.5" Glew (1989) corer.}
}
\description{
Calculate core area from an internal diameter
}
\examples{
pb210_core_area()

}
\references{
Glew, J.R. 1989. A new trigger mechanism for sediment samplers.
Journal of Paleolimnology, 2: 241–243. doi:10.1007/BF00195474.
}
