% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pb210_decay_constant}
\alias{pb210_decay_constant}
\title{Calculate a decay constant}
\usage{
pb210_decay_constant(half_life = 22.26)
}
\arguments{
\item{half_life}{A half-life value in years. For lead-210, this value is 22.26 years.}
}
\description{
Calculate a decay constant
}
\examples{
pb210_decay_constant()

}
