% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{pb210_excess}
\alias{pb210_excess}
\title{Calculate excess (unsupported) lead-210}
\usage{
pb210_excess(activity, background = 0)
}
\arguments{
\item{activity}{A vector of measured lead-210 specific activities (in Bq/kg) and
associated error. These can have \code{\link[errors:errors]{errors::errors()}}.}

\item{background}{A vector of estimated background
lead-210 specific activity (in Bq/kg) and associated error.}
}
\value{
A vector with \code{\link[errors:errors]{errors::errors()}} of the excess lead-210 specific activity.
}
\description{
Calculate excess (unsupported) lead-210
}
\examples{
core <- pb210_simulate_core(depth_step = rep(1, 30)) \%>\%
  pb210_simulate_counting()

pb210_excess(
  set_errors(
    core$pb210_specific_activity_estimate,
    core$pb210_specific_activity_sd
  ),
  background = 10
)

}
