% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_parameters.R
\name{pb210_excess}
\alias{pb210_excess}
\title{Calculate excess (unsupported) lead-210}
\usage{
pb210_excess(total_pb210, background_pb210 = 0,
  total_pb210_sd = NA_real_, background_pb210_sd = NA_real_)
}
\arguments{
\item{total_pb210, total_pb210_sd}{A vector of measured lead-210 specific activities (in Bq/kg) and
associated error. These can have \code{\link[errors:errors]{errors::errors()}}.}

\item{background_pb210, background_pb210_sd}{A vector of estimated background
lead-210 specific activity (in Bq/kg) and associated error.}
}
\value{
A vector with \code{\link[errors:errors]{errors::errors()}} of the excess lead-210 specific activity.
}
\description{
Calculate excess (unsupported) lead-210
}
\examples{
core <- pb210_simulate_core(depth_step = rep(1, 30)) \%>\%
  pb210_simulate_counting()

pb210_excess(
  core$pb210_specific_activity_estimate,
  background_pb210 = 10,
  core$pb210_specific_activity_sd
)

}
