% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pb210_fit_exponential}
\alias{pb210_fit_exponential}
\alias{pb210_fit_loglinear}
\alias{pb210_fit_exponential_manual}
\alias{pb210_fit_exponential_zero}
\alias{pb210_fit_exponential_constant}
\alias{predict.exponential_manual}
\alias{coef.exponential_manual}
\alias{predict.lm_loglinear}
\alias{coef.lm_loglinear}
\title{Fit an exponential model}
\usage{
pb210_fit_exponential(x, y, subset = NULL)

pb210_fit_loglinear(x, y, subset = NULL)

pb210_fit_exponential_manual(m, b)

pb210_fit_exponential_zero()

pb210_fit_exponential_constant(value)

\method{predict}{exponential_manual}(object, newdata, ...)

\method{coef}{exponential_manual}(object, ...)

\method{predict}{lm_loglinear}(object, newdata, ...)

\method{coef}{lm_loglinear}(object, ...)
}
\arguments{
\item{x}{An independent variable like depth or cumulative dry mass.}

\item{y}{A dependent variable that responds exponentially to \code{x}.}

\item{subset}{\code{NULL} to do no subsetting, a logical or numeric
vector to subset manually, or a function of two variables
(\code{x} and \code{y}) that returns \code{NULL}, a logical vector, or
a numeric vector that will be used to subset the data.
This is particularly useful in conjunction with
\code{\link[=finite_head]{finite_head()}} or \code{\link[=finite_tail]{finite_tail()}}.}

\item{m, b, value}{Directly specify coefficients for a manual fit.}

\item{object}{A model fit object.}

\item{newdata}{A tibble with a column \code{x}.}

\item{...}{Not used.}
}
\value{
A model object like that returned by \code{\link[stats:nls]{stats::nls()}}, with a
\code{\link[stats:predict]{stats::predict()}} method.
}
\description{
Fits the exponential model (\code{y ~ exp(m * x + b)}), estimating parameters \code{m}
and \code{b} using \code{\link[stats:nls]{stats::nls()}}. The log-linear version fits the model \code{log(y) ~ x}
using \code{\link[stats:lm]{stats::lm()}}, where the y-intercept is \code{b} and the slope of the
line is \code{m}. The log-linear version overestimates the importance of small
values but is widely used. NA and zero values are removed observation-wise
prior to fitting.
}
\examples{
fake_depth <- 0:10
fake_pb210 <- exp(5 - fake_depth) + rnorm(11, sd = 0.005)

fit_exp <- pb210_fit_exponential(fake_depth, fake_pb210)
fit_loglinear <- pb210_fit_loglinear(fake_depth, fake_pb210)

coefficients(fit_exp)
coefficients(fit_loglinear)

tibble::tibble(
  new_depth = 0:5,
  fitted_exp = predict(fit_exp, newdata = tibble::tibble(x = new_depth)),
  fitted_loglin = predict(fit_loglinear, newdata = tibble::tibble(x = new_depth))
)

}
