% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{pb210_inventory}
\alias{pb210_inventory}
\alias{pb210_inventory_calculator}
\alias{predict.inventory_calculator}
\title{Calculate cumulative lead-210 activity}
\usage{
pb210_inventory(
  cumulative_dry_mass,
  excess,
  model_top = max_finite(excess),
  model_bottom = pb210_fit_exponential(cumulative_dry_mass, excess)
)

pb210_inventory_calculator(
  model_top = ~max_finite(..2),
  model_bottom = ~pb210_fit_exponential(..1, ..2)
)

\method{predict}{inventory_calculator}(object, cumulative_dry_mass, excess, ...)
}
\arguments{
\item{cumulative_dry_mass}{The cumulative dry mass of the core (in kg), starting at the
surface sample and including all samples in the core.
These must be greater than 0 and in increasing order.}

\item{excess}{An excess (non-erosional) lead-210 specific activity (in Bq/kg)
for samples where this was measured, and NA where lead-210 was not measured. Use
\code{\link[errors:set_errors]{errors::set_errors()}} to use quadrature error propogation.}

\item{model_top}{A fit object, such as one generated by \code{\link[=pb210_fit_exponential]{pb210_fit_exponential()}} or a
constant specifying the surface \code{excess}. The choice of this value has
considerable impact on young dates.}

\item{model_bottom}{A fit object that will be used to model activities below
the last positive finite lead-210 activity. Use a constant to specify a custom
deep inventory of lead-210. Can also be a lambda-function of
\code{cumulative_dry_mass}, and \code{excess}.}

\item{object}{An inventory calculator generated by
\code{\link[=pb210_inventory_calculator]{pb210_inventory_calculator()}}.}

\item{...}{Unused.}
}
\value{
\code{\link[=pb210_inventory]{pb210_inventory()}} returns a vector with \code{\link[errors:errors]{errors::errors()}} of
cumulative lead-210 activities for each sample (in Bq) that can be passed
as the \code{inventory} to \code{\link[=pb210_crs]{pb210_crs()}}. \code{\link[=pb210_inventory_calculator]{pb210_inventory_calculator()}} returns
a fit object that can be used to calculate inventory given a
\code{cumulative_dry_mass} and an \code{excess}.
}
\description{
The cumulative content of lead-210 from the bottom of the core is the basis
for the constant rate of supply model, and is a required input to
\code{\link[=pb210_crs]{pb210_crs()}}. There are as many ways to calculate this as there are people
who interpret lead-210 activities. This function models the bottom (below
measured lead-210), middle (between measured values of lead-210), and top
(above measured lead-210) with separate models to accomodate the variety of
methods. By default, lead-210 activity is estimated for samples in which it
was not measured by an exponential fit of lead-210 activity vs. cumulative dry mass
(bottom), and repeating the maximum value (top). Inventory for all values is
estimated using a trapezoidal approximation.
}
\examples{
fake_mass <- 1:10
fake_pb210 <- exp(5 - fake_mass)
pb210_inventory(fake_mass, fake_pb210)

# compare with known inventory from integrating
# exp(5 - fake_mass) to +Inf
exp(-1 * fake_mass  + 5) / -(-1)

}
