% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{pb210_inventory}
\alias{pb210_inventory}
\alias{pb210_inventory_calculator}
\alias{predict.inventory_calculator}
\title{Calculate cumulative lead-210 activity}
\usage{
pb210_inventory(cumulative_dry_mass, excess_pb210,
  model_top = max(excess_pb210, na.rm = TRUE),
  model_bottom = pb210_fit_exponential(cumulative_dry_mass,
  excess_pb210))

pb210_inventory_calculator(model_top = ~max(..2, na.rm = TRUE),
  model_bottom = ~pb210_fit_exponential(..1, ..2))

\method{predict}{inventory_calculator}(object, cumulative_dry_mass,
  excess_pb210, ...)
}
\arguments{
\item{cumulative_dry_mass}{The cumulative dry mass of the core (in kg), starting at the
surface sample and including all samples in the core.
These must be greater than 0 and in increasing order.}

\item{excess_pb210}{An excess lead-210 activity for samples where this was
measured, and NA where lead-210 was not measured. NA values will be
estimated using \code{model_top}, \code{model_middle}, and \code{model_bottom}.}

\item{model_top}{A fit object or constant that will be used to model
activities above the first positive finite lead-210 activity. Use NULL for
the default (use the maximum lead-210 activity). Can also be a
lambda-function of \code{cumulative_dry_mass}, \code{excess_pb210}, and
\code{excess_pb210_sd}.}

\item{model_bottom}{A fit object that will be used to model activities below
the last positive finite lead-210 activity. This must be created using
\code{\link[=pb210_fit_exponential]{pb210_fit_exponential()}} or \code{\link[=pb210_fit_loglinear]{pb210_fit_loglinear()}} in that the fit's \code{m}
and \code{b} coefficients are used to calculate the integrated activity below
the last positive finite \code{excess_pb210}. Use a constant to specify a custom
deep inventory of lead-210. Use NULL for the default
(\code{\link[=pb210_fit_exponential]{pb210_fit_exponential()}}). Can also be a lambda-function of
\code{cumulative_dry_mass}, \code{excess_pb210}, and \code{excess_pb210_sd}.}

\item{object}{An inventory calculator generated by
\code{\link[=pb210_inventory_calculator]{pb210_inventory_calculator()}}.}

\item{...}{Unused.}
}
\value{
\code{\link[=pb210_inventory]{pb210_inventory()}} returns a vector with \code{\link[errors:errors]{errors::errors()}} of
cumulative lead-210 activities for each sample (in Bq) that can be passed
as the \code{inventory} to \code{\link[=pb210_crs]{pb210_crs()}}. \code{\link[=pb210_inventory_calculator]{pb210_inventory_calculator()}} returns
a fit object that can be used to calculate inventory given an
\code{cumulative_dry_mass}, \code{excess_pb210}, and \code{excess_pb210_sd}.
}
\description{
The cumulative content of lead-210 from the bottom of the core is the basis
for the constant rate of supply model, and is a required input to
\code{\link[=pb210_crs]{pb210_crs()}}. There are as many ways to calculate this as there are people
who interpret lead-210 activities. This function models the bottom (below
measured lead-210), middle (between measured values of lead-210), and top
(above measured lead-210) with separate models to accomodate the variety of
methods. By default, lead-210 activity is estimated for samples in which it
was not measured by an exponential fit of lead-210 activity vs. depth (top
and bottom), and by linear interpolation between values (middle).
}
\examples{
fake_mass <- 1:10
fake_pb210 <- exp(5 - fake_mass)
pb210_inventory(fake_mass, fake_pb210)

# compare with known inventory from integrating
# exp(5 - fake_mass) to +Inf
exp(-1 * fake_mass  + 5) / -(-1)

}
