% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pb210_mass_accumulation_constant}
\alias{pb210_mass_accumulation_constant}
\alias{pb210_mass_accumulation_rlnorm}
\title{Parameter generators for mass accumulation rates}
\usage{
pb210_mass_accumulation_constant(value = 0.15)

pb210_mass_accumulation_rlnorm(mean = 0.15, sd = 0.02)
}
\arguments{
\item{value, mean}{An accumulation rate, in kg / m2 / year.}

\item{sd}{For random (log)normal accumulation rates, the standard deviation}
}
\value{
A function of a single parameter, \code{age}, which is a (decreasing) vector of ages.
}
\description{
Parameter generators for mass accumulation rates
}
\examples{
age_compare <- tibble::tibble(
  ages = 150:0,
  constant = pb210_mass_accumulation_constant()(ages),
  rlnorm = pb210_mass_accumulation_rlnorm()(ages)
)

}
