% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{pb210_sample_norm}
\alias{pb210_sample_norm}
\title{Random samplers}
\usage{
pb210_sample_norm(x)
}
\arguments{
\item{x}{A vector of values, possibly with \code{\link[errors:errors]{errors::errors()}}.}
}
\value{
A (errorless) vector of randomly sampled values if \code{x} has
\code{\link[errors:errors]{errors::errors()}}, otherwise \code{x}, unchanged.
}
\description{
Random samplers
}
\examples{
pb210_sample_norm(set_errors(1:10, 1))
pb210_sample_norm(set_errors(1:10, 1))
pb210_sample_norm(1:10)

}
