% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pb210_simulate_accumulation}
\alias{pb210_simulate_accumulation}
\alias{pb210_simulate_core}
\alias{pb210_simulate_counting}
\title{Simulate sediment accumulation}
\usage{
pb210_simulate_accumulation(
  mass_accumulation = pb210_mass_accumulation_constant(),
  max_age = 500,
  time_step = 1,
  supply = pb210_supply_constant(),
  compressibility = pb210_compressibility_constant(),
  initial_density = pb210_density_constant(),
  decay_constant = pb210_decay_constant()
)

pb210_simulate_core(
  accumulation = pb210_simulate_accumulation(),
  depth_step = rep(0.5, 60),
  core_area = pb210_core_area()
)

pb210_simulate_counting(
  accumulation = pb210_simulate_core(),
  count_mass = 0.5/1000,
  count_time = lubridate::ddays(1)
)
}
\arguments{
\item{mass_accumulation}{A function such as that generated by \code{\link[=pb210_mass_accumulation_constant]{pb210_mass_accumulation_constant()}}.
Values returned by the function should be in kg / m^2^ / year.}

\item{max_age}{The maximum age of the simulation (years)}

\item{time_step}{The time to consider in each step (years)}

\item{supply}{A function such as that generated by \code{\link[=pb210_supply_constant]{pb210_supply_constant()}}.
The flux of lead-210 to the surface, in Bq/ m^2^ / year}

\item{compressibility}{A function such as that generated by \code{\link[=pb210_compressibility_constant]{pb210_compressibility_constant()}}.
Values returned by the function should be in 1 / Pa.}

\item{initial_density}{A function such as that generated by \code{\link[=pb210_density_constant]{pb210_density_constant()}}.
Values returned by the function should be in kg / m^3^.}

\item{decay_constant}{The decay constant to use (see \code{\link[=pb210_decay_constant]{pb210_decay_constant()}})}

\item{accumulation}{An age simulation, created by \code{\link[=pb210_simulate_accumulation]{pb210_simulate_accumulation()}}.}

\item{depth_step}{A vector of depth steps to consider, from the top to the bottom
of the core. There should be one element in the vector for each slice considered.}

\item{core_area}{The internal area of the core in m^2^.}

\item{count_mass}{The amount of mass each sample that was counted (kg). Can be
a vector to specify different masses for each sample.}

\item{count_time}{The time each sample was counted, as resolved by
\code{\link[lubridate:as.duration]{lubridate::as.duration()}}. Raw numbers are treated as a number of seconds, however
we reccomend using \code{\link[lubridate:ddays]{lubridate::ddays()}} and/or \code{\link[lubridate:dhours]{lubridate::dhours()}} to make
the units explicit.}
}
\value{
A tibble with columns:
\itemize{
\item \strong{age} (years): the mean age of the slice (weighted by mass)
\item \strong{depth} (cm): the midpoint depth of the slice
\item \strong{activity} (Bq/kg)
\item \strong{age_top}, \strong{age_bottom} (years)
\item \strong{depth_top}, \strong{depth_bottom} (cm)
\item \strong{slice_mass} (kg)
\item \strong{slice_density} (kg / m^3^)
}

The output of \code{pb210_simulate_counting()} also includes the columns:
\itemize{
\item \strong{activity_estimate} (Bq/kg): A plausible activity that might be
obtained by a lead-210 counting method, sampled using \code{\link[stats:rpois]{stats::rpois()}}.
\item \strong{activity_se} (Bq/kg): The standard error (estimate of the standard
deviation) of \strong{activity_estimate}. For the expected error,
use \code{\link[=pb210_error_from_activity]{pb210_error_from_activity()}} on
}
}
\description{
\code{pb210_simulate_accumulation()} simulates the accumulation of lead-210 over
time, returning a data frame with one row per \code{time_step}, whereas
\code{pb210_simulate_core()} returns a data frame with one row per \code{depth_step}
(like extruding a core). These functions were created to test the
\code{\link[=pb210_crs]{pb210_crs()}} and \code{\link[=pb210_cic]{pb210_cic()}} functions, and to create
the lead-210 theory vignette. Ages are in years, depths are in cm,
masses are in kg, densities are in kg / m^3^, and specific activities
are in Bq / kg. \code{\link[=pb210_simulate_counting]{pb210_simulate_counting()}} simulates putting a subsample
of each slice on a counter for a specified amount of time. This allows
assigning an error to the known lead-210 activity, as well as provide
more realistic data for simulations. See \code{\link[=pb210_error_from_activity]{pb210_error_from_activity()}}
for details.
}
\examples{
# 1 row per year
accumulation <- pb210_simulate_accumulation()
# 1 row per 0.5 cm
core <- pb210_simulate_core(accumulation)
counted_core <- pb210_simulate_counting(core)

}
