% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pb210_simulate_accumulation}
\alias{pb210_simulate_accumulation}
\title{Simulate sediment accumulation}
\usage{

  pb210_simulate_accumulation(mass_accumulation = pb210_mass_accumulation_constant(),
  max_age = 500, time_step = 1, supply = pb210_supply_constant(),
  compressibility = pb210_compressibility_constant(),
  initial_density = pb210_density_constant(),
  decay_constant = pb210_decay_constant())
}
\arguments{
\item{mass_accumulation}{A function such as that generated by \code{\link[=pb210_mass_accumulation_constant]{pb210_mass_accumulation_constant()}}.
Values returned by the function should be in kg / m^2^ / year.}

\item{max_age}{The maximum age of the simulation (years)}

\item{time_step}{The time to consider in each step (years)}

\item{supply}{A function such as that generated by \code{\link[=pb210_supply_constant]{pb210_supply_constant()}}.
The flux of lead-210 to the surface, in Bq/ m^2^ / year}

\item{compressibility}{A function such as that generated by \code{\link[=pb210_compressibility_constant]{pb210_compressibility_constant()}}.
Values returned by the function should be in 1 / Pa.}

\item{initial_density}{A function such as that generated by \code{\link[=pb210_density_constant]{pb210_density_constant()}}.
Values returned by the function should be in kg / m^3^.}

\item{decay_constant}{The decay constant to use (see \code{\link[=pb210_decay_constant]{pb210_decay_constant()}})}
}
\value{
A tibble with components ...
}
\description{
Simulate sediment accumulation
}
\examples{
pb210_simulate_accumulation()

}
