% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pb210_simulate_core}
\alias{pb210_simulate_core}
\title{Simulate a core of an accumulation}
\usage{
pb210_simulate_core(accumulation = pb210_simulate_accumulation(),
  depth_step = rep(0.5, 60), core_area = pb210_core_area())
}
\arguments{
\item{accumulation}{An age simulation, created by \code{\link[=pb210_simulate_accumulation]{pb210_simulate_accumulation()}}.}

\item{depth_step}{A vector of depth steps to consider, from the top to the bottom
of the core. There should be one element in the vector for each slice considered.}

\item{core_area}{The internal area of the core in m^2^.}
}
\value{
A tibble with components ...
}
\description{
Simulate a core of an accumulation
}
\examples{
pb210_simulate_core()

}
